+++ PDKit/Commands/ChangeHormoneCommand.swift
@@ -1,5 +1,5 @@
 //
-//  PDCommand.swift
+//  ChangeHormoneCommand.swift
 //  PDKit
 //
 //  Created by Juliya Smith on 10/31/20.