+++ PDKit/Protocols/UI/TabReflective.swift
@@ -1,5 +1,5 @@
 //
-//  PDTabReflective.swift
+//  TabReflective.swift
 //  PatchDay
 //
 //  Created by Juliya Smith on 8/10/19.
@@ -9,10 +9,22 @@
 import UIKit
 
 public protocol TabReflective {
-    var hormonesVC: UIViewController? { get }
-    var pillsVC: UIViewController? { get }
-    var sitesVC: UIViewController? { get }
+
+    /// The Hormones View Controller.
+    var hormonesViewController: UIViewController? { get }
+
+    /// The Pills View Controller.
+    var pillsViewController: UIViewController? { get }
+
+    /// The Sites View Controller.
+    var sitesViewController: UIViewController? { get }
+
+    /// Sets the Hormones and Pills tabs to their respective expired counts.
     func reflect()
+
+    /// Sets the Hormones tab badge value to the total expired hormones.
     func reflectHormones()
+
+    /// Sets the Pills tab badge value to the total expired pills.
     func reflectPills()
 }