+++ PatchDay/Hormones/Details/HormoneDatePicker.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class HormoneDatePicker: UIDatePicker {
 
-    var setDateSideEffect: () -> () = {}
+    var setDateSideEffect: () -> Void = {}
 
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         super.touchesBegan(touches, with: event)